Stable version of the firmware
Firmware is not final, but it now supports the remote volume control (with ZERO pops!), so it's time to check in a version that basically works.  There is more polishing to do here, certainly, but this version basically works, and power consumption is pretty low on the V2.2 board (about 4.4mA).  Uses only about 13.6% of the available space (1117/8192 bytes).