Tweaked low battery warning
Low battery warning starts flashing the LED at about 7.2V now, and the LED will be on less and less down to about 6.0V when it's on only 1/10 of the time.  At that point, battery life will be a handful of hours (exact number depends on the battery type).  Note that USB rechargeable 9V batteries do not give any warning before their output drops suddenly to nearly zero, so the flashing warning applies to non-rechargeable Alkaline and Lithium batteries only.