Update firmware to handle RVC better
The RVC curve now has 5 sections:  2 guard bands around each end (for tolerances), a MUTE section (-64dB = mute), and two sloped sections (gentle for the first 3/4, aggressive for the last 1/4.  This is very much like a Hilton curve, except it has the aggressive attenuation down to mute at the end.